+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features:**
 * Added `preserveSearch` option, enables preservation the search fragment when only the hash changes.
+* Added the ability to change arbitrary options. Use carefully!
 
 **Bugfixes:**
 * Fixed a quirk in `push` and `replace` keeping empty hash and search when called with a Location but not when called with a string.